Fix issues reported by SonarCloud

+++ security/unittest/doxmtest.cpp
@@ -61,8 +61,8 @@ IsEqual(const oc_sec_doxm_t &lhs, const oc_sec_doxm_t &rhs,
     if (lhs.num_oxms != rhs.num_oxms) {
       return false;
     }
-    for (int i = 0;
-         i < std::min<int>(lhs.num_oxms, static_cast<int>(std::size(lhs.oxms)));
+    for (int i = 0; i < std::min<size_t>(static_cast<size_t>(lhs.num_oxms),
+                                         std::size(lhs.oxms));
          ++i) {
       if (lhs.oxms[i] != rhs.oxms[i]) {
         return false;
